refactored function in the flex component
